Refactor Codex Connector review churn diagnostics
