+++ src/codex-connector-review-policy.test.ts
@@ -1,17 +1,19 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 import {
-  buildCodexConnectorP2P3PolicyDiagnostic,
-  buildCodexConnectorPolicyBlockDiagnostic,
   buildCodexConnectorReviewChurnDiagnostic,
   buildCodexConnectorReviewChurnProgressSummary,
-  buildReviewPolicyInput,
   clusterConfiguredBotReviewThreads,
   compareCodexConnectorReviewChurnProgress,
+  formatCodexConnectorReviewChurnDiagnostic,
+} from "./codex-connector-review-churn";
+import {
+  buildCodexConnectorP2P3PolicyDiagnostic,
+  buildCodexConnectorPolicyBlockDiagnostic,
+  buildReviewPolicyInput,
   codexConnectorMustFixReviewThreads,
   codexConnectorStaleReviewCommitThreads,
   evaluateCodexConnectorConvergencePolicy,
-  formatCodexConnectorReviewChurnDiagnostic,
 } from "./codex-connector-review-policy";
 import type { GitHubPullRequest, ReviewThread } from "./core/types";
 import { createConfig, createReviewThread } from "./turn-execution-test-helpers";
