+++ src/recovery-blocked-issue-reconciliation.ts
@@ -44,6 +44,8 @@ import {
 import {
   buildCodexConnectorReviewChurnDiagnostic,
   buildCodexConnectorReviewChurnProgressSummary,
+} from "./codex-connector-review-churn";
+import {
   codexConnectorMustFixReviewThreads,
   hasCodexConnectorFindingReviewComment,
   latestCodexConnectorReviewCommentFingerprint,
