+++ src/supervisor/stale-review-bot-remediation.ts
@@ -8,6 +8,8 @@ import {
 } from "../review-handling";
 import {
   clusterConfiguredBotReviewThreads,
+} from "../codex-connector-review-churn";
+import {
   codexConnectorMustFixReviewThreads,
   commitShasEqualForComparison,
   evaluateCodexConnectorConvergencePolicy,
