+++ src/turn-execution-orchestration.ts
@@ -21,6 +21,8 @@ import {
 } from "./review-handling";
 import {
   buildCodexConnectorReviewChurnDiagnostic,
+} from "./codex-connector-review-churn";
+import {
   codexConnectorMustFixReviewThreads,
   latestCodexConnectorReviewCommentFingerprint,
 } from "./codex-connector-review-policy";
